% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEnrichMethod.R
\name{listEnrichMethod}
\alias{listEnrichMethod}
\title{List of enrichment methods}
\usage{
listEnrichMethod()
}
\value{
A character vector containing the names of supported enrichment methods:
\itemize{
\item GO - Gene Ontology
\item KEGG - Kyoto Encyclopedia of Genes and Genomes
\item MKEGG - Module KEGG
\item WikiPathways - WikiPathways database
\item Reactome - Reactome pathway database
\item MsigDB - Molecular Signatures Database
\item DO - Disease Ontology
\item CGN - Cancer Gene Network
\item DisGeNET - DisGeNET database
\item CellMarker - CellMarker database
\item CMAP - Connectivity Map (drug signatures)
}
}
\description{
Returns a vector of available enrichment analysis methods supported by BioEnricher.
These methods can be used with the lzq_ORA and lzq_GSEA functions.
}
\examples{
\dontrun{
# List all available enrichment methods
methods <- listEnrichMethod()
print(methods)

# Use in enrichment analysis
ora_result <- lzq_ORA(genes = my_genes, enrich.type = "GO")
gsea_result <- lzq_GSEA(genes = my_gene_scores, enrich.type = "KEGG")
}

}
