% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzq_getEF.R
\name{lzq_getEF}
\alias{lzq_getEF}
\title{Get enrichment factor from enrichResult}
\usage{
lzq_getEF(res)
}
\arguments{
\item{res}{An enrichResult object from ORA or GSEA analysis}
}
\value{
The same enrichResult object with an additional "EnrichmentFactor" column
}
\description{
Calculates and adds the enrichment factor to an enrichResult object.
The enrichment factor is the ratio of gene ratio to background ratio,
indicating how much more enriched a pathway is compared to random expectation.
}
\examples{
\dontrun{
# Add enrichment factor to ORA results
ora_results <- lzq_ORA(genes = my_genes, enrich.type = "GO")
ora_with_ef <- lzq_getEF(ora_results)
}

}
