% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzq_GSEA.barplot1.R
\name{lzq_GSEA.barplot1}
\alias{lzq_GSEA.barplot1}
\title{Enrichment barplot for positive or negative GSEA results}
\usage{
lzq_GSEA.barplot1(
  enrich.obj,
  type = "Positive",
  show.term.num = 15,
  Selct.P = "FDR",
  cutoff.P = 0.05,
  colors = c("#003c30", "#01665e", "#35978f", "#80cdc1", "#c7eae5", "#f6e8c3", "#dfc27d",
    "#bf812d", "#8c510a", "#543005"),
  add.bar.border = T,
  bar.width = 0.6,
  y.label.position = "right",
  title = NULL,
  legend.position = "right",
  theme.plot = theme_bw(base_rect_size = 1.5),
  use.Chinese = F,
  appid = "20231122001888718",
  key = "5GpDqe8F3pmXfnOkEKGQ"
)
}
\arguments{
\item{enrich.obj}{A gseaResult object from GSEA analysis}

\item{type}{Type of enrichment to show ("Positive" or "Negative"). Default is "Positive".}

\item{show.term.num}{Number of top terms to display. Default is 15.}

\item{Selct.P}{Type of p-value to use for filtering ("FDR", "pvalue"). Default is "FDR".}

\item{cutoff.P}{P-value cutoff for significance. Default is 0.05.}

\item{colors}{Vector of colors for the gradient. Default is a green-brown palette.}

\item{add.bar.border}{Logical, whether to add borders to bars. Default is TRUE.}

\item{bar.width}{Width of bars (0-1). Default is 0.6.}

\item{y.label.position}{Position of y-axis labels ("right", "left"). Default is "right".}

\item{title}{Plot title. Default is NULL.}

\item{legend.position}{Position of legend ("right", "left", "top", "bottom"). Default is "right".}

\item{theme.plot}{ggplot2 theme to apply. Default is theme_bw(base_rect_size = 1.5).}

\item{use.Chinese}{Logical, whether to translate terms to Chinese. Default is FALSE.}

\item{appid}{Baidu translation API ID (if use.Chinese = TRUE). Default is "20231122001888718".}

\item{key}{Baidu translation API key (if use.Chinese = TRUE). Default is "5GpDqe8F3pmXfnOkEKGQ".}
}
\value{
A ggplot2 object representing the GSEA barplot.
}
\description{
Creates a horizontal barplot visualization for Gene Set Enrichment Analysis (GSEA) results,
showing either positively or negatively enriched gene sets.
}
\examples{
\dontrun{
# Plot positive GSEA results
gsea_plot <- lzq_GSEA.barplot1(enrich.obj = gsea_results, type = "Positive")

# Plot negative GSEA results
gsea_plot <- lzq_GSEA.barplot1(enrich.obj = gsea_results, 
                              type = "Negative",
                              show.term.num = 10)
}

}
