% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzq_ORA.R
\name{lzq_ORA}
\alias{lzq_ORA}
\title{Over-representative analysis}
\usage{
lzq_ORA(
  genes,
  background.genes = NULL,
  gene.type = "SYMBOL",
  enrich.type,
  organism = "Human",
  GO.ont = "BP",
  GO.simplify = T,
  KEGG.use.internal.data = F,
  MsigDB.category = "H",
  CMAP.min.Geneset.Size = 3,
  pvalue.cutoff = 0.05,
  qvalue.cutoff = 0.05,
  padjust.method = "BH",
  min.Geneset.Size = 10,
  max.Geneset.Size = 1000
)
}
\arguments{
\item{genes}{Vector of genes for enrichment analysis}

\item{background.genes}{Vector of background genes. If NULL, uses all genes in the database. Default is NULL.}

\item{gene.type}{Type of gene identifiers ("SYMBOL", "ENTREZID", etc.). Default is "SYMBOL".}

\item{enrich.type}{Type of enrichment analysis to perform (e.g., "GO", "KEGG", "WikiPathways")}

\item{organism}{Organism for analysis ("Human", "Mouse", etc.). Default is "Human".}

\item{GO.ont}{GO ontology type ("BP", "CC", "MF"). Default is "BP".}

\item{GO.simplify}{Logical, whether to simplify GO results. Default is TRUE.}

\item{KEGG.use.internal.data}{Logical, whether to use internal KEGG data. Default is FALSE.}

\item{MsigDB.category}{MsigDB category to use. Default is "H".}

\item{CMAP.min.Geneset.Size}{Minimum gene set size for CMAP analysis. Default is 3.}

\item{pvalue.cutoff}{P-value cutoff for significance. Default is 0.05.}

\item{qvalue.cutoff}{Q-value cutoff for significance. Default is 0.05.}

\item{padjust.method}{Method for p-value adjustment. Default is "BH".}

\item{min.Geneset.Size}{Minimum gene set size. Default is 10.}

\item{max.Geneset.Size}{Maximum gene set size. Default is 1000.}
}
\value{
An enrichResult object containing the over-representation analysis results.
}
\description{
Performs over-representation analysis (ORA) for gene sets using various databases
including GO, KEGG, WikiPathways, Reactome, MsigDB, and others.
}
\examples{
\dontrun{
# GO enrichment analysis
ora_results <- lzq_ORA(genes = my_genes, enrich.type = "GO")

# KEGG pathway analysis
kegg_results <- lzq_ORA(genes = my_genes, enrich.type = "KEGG")

# With custom background
ora_custom <- lzq_ORA(genes = my_genes, 
                     background.genes = background_genes,
                     enrich.type = "GO")
}

}
